@@ -1,13 +1,16 @@
 buildscript {
     repositories {
-        jcenter()
         google()
+        mavenCentral()
+        maven { url "https://jitpack.io" }
+        maven { url "https://plugins.gradle.org/m2/" }
+        gradlePluginPortal()
     }
 
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.1.3'
-        classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.1'
-        classpath 'com.github.dcendents:android-maven-gradle-plugin:2.1'
+        classpath 'com.android.tools.build:gradle:8.5.1'
+        // Removed bintray plugin as it's deprecated and causing dependency issues
+        // classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.1'
     }
 }
 
@@ -26,7 +29,7 @@ ext {
     siteUrl = 'https://github.com/barteksc/PdfiumAndroid'
     gitUrl = 'https://github.com/barteksc/PdfiumAndroid.git'
 
-    libraryVersion = '1.9.0'
+    libraryVersion = '2.0.0'
 
     developerId = 'barteksc'
     developerName = 'Bartosz Schiller'
@@ -38,13 +41,20 @@ ext {
 }
 
 android {
-    compileSdkVersion 26
+    namespace 'com.shockwave.pdfium'
+    compileSdk  35
 
     defaultConfig {
-        minSdkVersion 14
-        targetSdkVersion 26
-        versionCode 1
-        versionName "1.9.0"
+        minSdkVersion 26
+        targetSdk 35
+        versionCode 2
+        versionName "2.0.0"
+        buildConfigField "String", "VERSION_NAME", "\"${versionName}\""
+        /*
+        // Support for 16KB page size devices (Android 14+)
+        ndk {
+            abiFilters 'arm64-v8a', 'armeabi-v7a', 'x86', 'x86_64'
+        }*/
     }
     buildTypes {
         release {
@@ -53,23 +63,52 @@ android {
         }
     }
 
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
+    }
+
     sourceSets{
         main {
             jni.srcDirs = []
-            jniLibs.srcDir 'src/main/libs'
+//            jniLibs.srcDir 'src/main/libs'
+            jniLibs.srcDir 'src/main/jni/lib'
+        }
+    }
+
+    buildFeatures {
+        viewBinding true
+        buildConfig true
+    }
+
+    ndkVersion '28.2.13676358'
+    // Configure for 16KB page size support
+    packagingOptions {
+        jniLibs {
+            useLegacyPackaging = false
+        }
+    }
+
+    // Configure AAR filename to include version
+    libraryVariants.all { variant ->
+        variant.outputs.all {
+            outputFileName = "PdfiumAndroid-${defaultConfig.versionName}-${variant.buildType.name}.aar"
         }
     }
 }
 
 repositories {
     google()
-    jcenter()
+//    jcenter()
+    mavenCentral()
+    maven { url "https://jitpack.io" }
+//    maven { url "https://maven.aliyun.com/repository/jcenter" }
 }
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:support-v4:26.1.0'
+    implementation 'androidx.collection:collection:1.3.0'
 }
 
-apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/installv1.gradle'
-apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/bintrayv1.gradle'
+//apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/installv1.gradle'
+//apply from: 'https://raw.githubusercontent.com/nuuneoi/JCenter/master/bintrayv1.gradle'
